P06-Full/python/requirements.txt
web3==6.11.3
eth-account==0.10.0
flask==3.0.0
requests==2.31.0
python-dotenv==1.0.0
eth-abi==4.2.1
pydantic==2.5.0
colorama==0.4.6
websockets==12.0

